clean up image dimensions files
I left some sort of mess involving the version with "dont_know_if_needed" in the name. I think it was a temporary copy that I made for some reason. So for now I'm keeping it under this name. But at some point in the future, it should probably just be deleted unless some use for it is discovered.